공동방 생성, apporve 생성, 승낙, 거절, 공동방 생성시 멤버들에게 공동todo 생성까지 완

+++ DDIS_Project/src/main/java/com/DDIS/shareTodo/Command/domain/aggregate/Entity/Posts.java
@@ -15,7 +15,7 @@ public class Posts {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "post_num")
-    private int postNum;
+    private Long postNum;
 
     @Column(name = "post_title")
     private String postTitle;
